Fixed incorrect type in function declaration.

+++ Source/File.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <cstdint>
+
 namespace File {
-    void* Load(const char* filename, size_t* size);
+    void* Load(const char* filename, uint64_t* size);
     void Free(void* ptr);
 };
